Add Pro S RGB revision A schematic (US Layout)
Revision A schematic covers the following circuits:
- Root (pros_rgb.kicad_sch):
  - HT32F1654
  - SPI flash
  - USB connector
  - 5V to 3V3
  - MBIA 3V3
  - Debug connectors (SWD, SEL1, SEL2)
  - RGB LED circuits (led_circuits.kicad_sch):
    - MBIA RGB channel drivers
    - LED power drivers
    - LED groupings (by common RGB pins) (led_groupN.kicad_sch)
  - Key Switch circuits (key_switches.kicad_sch):
    - Key switch groupings (same groups as LEDs) (key_groupN.kicad_sch)

After spending a few days tracing the schematic, I felt this was a good
point to "save" my progress even if it is missing some details. There
are some components I have not been able to identify and were left as ?s
or filled in with placeholder symbols (ones in SMD_components.kicad_sym).
For REing and writing firmware, this is pretty sufficient (it is not
like I plan to recreate the PCB).